+++ app/templates/admin/admin_panel.html
